requirements.txt
numpy==1.18.0
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
idna==2.8
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
mysql-connector-python==8.0.15
protobuf==3.6.1
pylint==2.2.2
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.24.1
wrapt==1.11.1
Flask==1.0.2
jira==2.0.0
python-pptx==0.6.18
Flask-RESTful==0.3.7
waitress==1.3.0
toolz==0.10.0
pandas==0.25.3
opencv-python==4.3.0.36
WTForms==2.3.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3